+++ CHANGELOG.md
@@ -4,7 +4,9 @@ SQLRunner version is versioned following [semver](https://semver.org). All new
 features and stuff go under `Added`, All fixes go under `Fixed`, deprecated
 and removed features go under `Deprecated` and `Removed` respectively.
 
-## [Unreleased]
+## Unreleased
+
+## [0.1.1] - 2021-03-15
 
 ### Fixed